src/main/webapp/app/entities/owner-my-suffix/delete/owner-my-suffix-delete-dialog.component.spec.ts
jest.mock('@ng-bootstrap/ng-bootstrap');

import { ComponentFixture, TestBed, inject, fakeAsync, tick } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { of } from 'rxjs';
import { NgbActiveModal } from '@ng-bootstrap/ng-bootstrap';

import { OwnerMySuffixService } from '../service/owner-my-suffix.service';

import { OwnerMySuffixDeleteDialogComponent } from './owner-my-suffix-delete-dialog.component';

describe('Component Tests', () => {
  describe('OwnerMySuffix Management Delete Component', () => {
    let comp: OwnerMySuffixDeleteDialogComponent;
    let fixture: ComponentFixture<OwnerMySuffixDeleteDialogComponent>;
    let service: OwnerMySuffixService;
    let mockActiveModal: NgbActiveModal;

    beforeEach(() => {
      TestBed.configureTestingModule({
        imports: [HttpClientTestingModule],
        declarations: [OwnerMySuffixDeleteDialogComponent],
        providers: [NgbActiveModal],
      })
        .overrideTemplate(OwnerMySuffixDeleteDialogComponent, '')
        .compileComponents();
      fixture = TestBed.createComponent(OwnerMySuffixDeleteDialogComponent);
      comp = fixture.componentInstance;
      service = TestBed.inject(OwnerMySuffixService);
      mockActiveModal = TestBed.inject(NgbActiveModal);
    });

    describe('confirmDelete', () => {
      it('Should call delete service on confirmDelete', inject(
        [],
        fakeAsync(() => {
          // GIVEN
          jest.spyOn(service, 'delete').mockReturnValue(of(new HttpResponse({})));

          // WHEN
          comp.confirmDelete(123);
          tick();

          // THEN
          expect(service.delete).toHaveBeenCalledWith(123);
          expect(mockActiveModal.close).toHaveBeenCalledWith('deleted');
        })
      ));

      it('Should not call delete service on clear', () => {
        // GIVEN
        jest.spyOn(service, 'delete');

        // WHEN
        comp.cancel();

        // THEN
        expect(service.delete).not.toHaveBeenCalled();
        expect(mockActiveModal.close).not.toHaveBeenCalled();
        expect(mockActiveModal.dismiss).toHaveBeenCalled();
      });
    });
  });
});
